Oy2 16334 - Update Waiver read-only form: Temporary Extension details (#818)
* update temp ext form config to display appropriate elements

* refactor detail nav

* refactored each detail view section component into its own file

* show additional info section on detail page without left nav
