+++ services/common/routes.js
@@ -62,6 +62,6 @@ export const TYPE_TO_DETAIL_ROUTE = {
   [ONEMAC_TYPE.WAIVER_BASE]: ONEMAC_ROUTES.BASE_WAIVER_DETAIL,
   [ONEMAC_TYPE.WAIVER_RENEWAL]: "",
   [ONEMAC_TYPE.WAIVER_AMENDMENT]: ONEMAC_ROUTES.WAIVER_AMENDMENT_DETAIL,
-  [ONEMAC_TYPE.WAIVER_EXTENSION]: "",
+  [ONEMAC_TYPE.WAIVER_EXTENSION]: ONEMAC_ROUTES.TEMPORARY_EXTENSION_DETAIL,
   [ONEMAC_TYPE.WAIVER_APP_K]: "",
 };
