+++ services/ui-src/src/page/base-waiver/BaseWaiverDetail.tsx
@@ -2,13 +2,15 @@ import React, { FC } from "react";
 import DetailView from "../DetailView";
 import {
   OneMACDetail,
-  DetailViewTab,
   defaultDetail,
   typeDefault,
   waiverAuthorityDefault,
   territoryDefault,
   submissionDateDefault,
   proposedEffectiveDateDefault,
+  defaultPackageOverviewNavItems,
+  defaultPackageOverviewLabel,
+  tempExtensionSectionNavItem,
 } from "../DetailViewDefaults";
 import { ROUTES, baseWaiver } from "cmscommonlib";
 
@@ -17,24 +19,8 @@ export const baseWaiverDetail: OneMACDetail = {
   ...baseWaiver,
   navItems: [
     {
-      label: "Package Overview",
-      items: [
-        {
-          id: DetailViewTab.DETAIL,
-          label: "Package Details",
-          url: `#${DetailViewTab.DETAIL}`,
-        },
-        {
-          id: DetailViewTab.ADDITIONAL,
-          label: "Additional Information",
-          url: `#${DetailViewTab.ADDITIONAL}`,
-        },
-        {
-          id: DetailViewTab.EXTENSION,
-          label: "Temporary Extension",
-          url: `#${DetailViewTab.EXTENSION}`,
-        },
-      ],
+      label: defaultPackageOverviewLabel,
+      items: [...defaultPackageOverviewNavItems, tempExtensionSectionNavItem],
     },
   ],
   detailSection: [
