+++ services/ui-src/src/page/temporary-extension/TemporaryExtensionDetail.tsx
@@ -1,14 +1,30 @@
 import React, { FC } from "react";
 import DetailView from "../DetailView";
-import { OneMACDetail, defaultDetail } from "../DetailViewDefaults";
+import {
+  OneMACDetail,
+  defaultDetail,
+  typeDefault,
+  submissionDateDefault,
+  AttributeDetail,
+} from "../DetailViewDefaults";
 import { Workflow, waiverTemporaryExtension } from "cmscommonlib";
+
+const parentIdDetail: AttributeDetail = {
+  heading: "Parent Waiver Number",
+  fieldName: "parentId",
+  default: null,
+};
 
 export const waiverTemporaryExtensionDetail: OneMACDetail = {
   ...defaultDetail,
   ...waiverTemporaryExtension,
+  detailHeader: "Temporary Extension",
   navItems: [],
+  usesVerticalNav: false,
   actionsByStatus: Workflow.waiverExtensionActionsByStatus,
+  show90thDayInfo: false,
   raiLink: "",
+  detailSection: [parentIdDetail, typeDefault, submissionDateDefault],
 };
 
 const WaiverTemporaryExtensionDetail: FC = () => {
